Update CodeMirror libraries from v3.19.1 to v5.21.0
These were incredibly out of date.

We may want to -- at some point -- migrate them out of Custom CSS and
into general scripts and styles, and add in PHP and Javascript support
as well, to integrate to the admin theme and plugin editor.

Or, alternately, run multiple files with dependencies.