Sync: Full Sync on update on MU
Do a full sync in a job if we are doing a full sync because of an
update to the plugin. Which send the full sync data such as options,
callbacks, network_options and constants for each of the sites on the
network.

This is supposed to reduce the load when updating Jetpack to a new
version.